Simplified pipeline fixture
