+++ tests/pipeline/test_codab.py
@@ -2,11 +2,9 @@
 from pathlib import Path
 
 import pytest
+from conftest import FAKE_AA_DATA_DIR, ISO3
 
 from aatoolbox.datasources.codab import _MODULE_BASENAME
-from aatoolbox.utils.io import parse_yaml
-from tests.conftest import FAKE_AA_DATA_DIR
-from tests.pipeline.conftest import CONFIG_FILE, ISO3
 
 
 @pytest.fixture(autouse=True)
@@ -21,9 +19,8 @@ def gpd_read_file(mocker):
     return mocker.patch("aatoolbox.datasources.codab.gpd.read_file")
 
 
-def test_codab_download(pipeline_caller, downloader):
+def test_codab_download(pipeline, downloader):
     """Test that get_codab calls the HDX API to download."""
-    pipeline = pipeline_caller()
     pipeline.get_codab(admin_level=2)
     downloader.assert_called_with(
         hdx_address=pipeline._config.codab.hdx_address,
@@ -34,9 +31,8 @@ def test_codab_download(pipeline_caller, downloader):
     )
 
 
-def test_codab_get_admin_level(pipeline_caller, gpd_read_file):
+def test_codab_get_admin_level(pipeline, gpd_read_file):
     """Test that get_codab retrieves expected file and layer name."""
-    pipeline = pipeline_caller()
     admin_level = 2
     expected_layer_name = pipeline._config.codab.layer_base_name.format(
         admin_level=admin_level
@@ -50,22 +46,18 @@ def test_codab_get_admin_level(pipeline_caller, gpd_read_file):
     )
 
 
-def test_codab_too_high_admin_level(pipeline_caller):
+def test_codab_too_high_admin_level(pipeline):
     """Test raised error when too high admin level requested."""
     with pytest.raises(AttributeError):
-        pipeline_caller().get_codab(admin_level=10)
+        pipeline.get_codab(admin_level=10)
 
 
-def test_codab_custom(pipeline_caller, gpd_read_file):
+def test_codab_custom(pipeline, gpd_read_file):
     """Test that get_codab_custom retrieves expected file and layer name."""
-    config_dict = parse_yaml(CONFIG_FILE)
-    custom_layer_name_list = ["custom_layer_A", "custom_layer_B"]
-    config_dict["codab"]["custom_layer_names"] = custom_layer_name_list
     custom_layer_number = 1
-
-    pipeline_caller(config_dict=config_dict).get_codab_custom(
-        custom_layer_number
-    )
+    custom_layer_name_list = ["custom_layer_A", "custom_layer_B"]
+    pipeline._config.codab.custom_layer_names = custom_layer_name_list
+    pipeline.get_codab_custom(custom_layer_number)
     gpd_read_file.assert_called_with(
         f"zip:///{FAKE_AA_DATA_DIR}/public/raw/{ISO3}/{_MODULE_BASENAME}/"
         f"{ISO3}_{_MODULE_BASENAME}.shp.zip/"
@@ -73,7 +65,7 @@ def test_codab_custom(pipeline_caller, gpd_read_file):
     )
 
 
-def test_codab_custom_missing(pipeline_caller, gpd_read_file):
+def test_codab_custom_missing(pipeline, gpd_read_file):
     """Test raised error when custom COD AB missing."""
     with pytest.raises(AttributeError):
-        pipeline_caller().get_codab_custom(0)
+        pipeline.get_codab_custom(0)
